AU-1038 Add base class for profile forms
